refactor: move Prisma client initialization to a separate module
