+++ src/app/boards/[slug]/page.tsx
@@ -1,4 +1,4 @@
-import prisma from "@/db";
+import { prisma } from "@/lib/prisma";
 
 export default async function BoardPage({
   params,
