+++ src/app/boards/page.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import prisma from "@/db";
+import { prisma } from "@/lib/prisma";
 
 export const dynamic = "force-dynamic";
