--- src/db.ts
@@ -1,9 +1,0 @@
-import { PrismaClient } from "@prisma/client";
-
-const globalForPrisma = global as unknown as { prisma: PrismaClient };
-
-const client = globalForPrisma.prisma || new PrismaClient();
-
-export default client;
-
-if (process.env.NODE_ENV !== "production") globalForPrisma.prisma = client;
