+++ src/lib/prisma.ts
@@ -1,0 +1,9 @@
+import { PrismaClient } from "@prisma/client";
+
+const prismaGlobal = globalThis as typeof globalThis & {
+  prisma?: PrismaClient;
+};
+
+export const prisma = prismaGlobal.prisma || new PrismaClient();
+
+if (process.env.NODE_ENV !== "production") prismaGlobal.prisma = prisma;
